chore(core): move project-json plugin to src
