+++ packages/nx/src/generators/utils/project-configuration.ts
@@ -4,7 +4,7 @@ import {
   buildProjectConfigurationFromPackageJson,
   getGlobPatternsFromPackageManagerWorkspaces,
 } from '../../../plugins/package-json-workspaces';
-import { buildProjectFromProjectJson } from '../../../plugins/project-json';
+import { buildProjectFromProjectJson } from '../../plugins/project-json/build-nodes/project-json';
 import { renamePropertyWithStableKeys } from '../../adapter/angular-json';
 import {
   ProjectConfiguration,
