+++ packages/nx/src/plugins/project-json/build-nodes/project-json.ts
@@ -4,33 +4,32 @@ import { existsSync } from 'node:fs';
 import {
   ProjectConfiguration,
   TargetConfiguration,
-} from '../src/config/workspace-json-project-json';
-import { toProjectName } from '../src/config/workspaces';
-import { readJsonFile } from '../src/utils/fileutils';
-import { NxPluginV2 } from '../src/utils/nx-plugin';
+} from '../../../config/workspace-json-project-json';
+import { toProjectName } from '../../../config/workspaces';
+import { readJsonFile } from '../../../utils/fileutils';
+import { NxPluginV2 } from '../../../utils/nx-plugin';
 import {
   PackageJson,
   readTargetsFromPackageJson,
-} from '../src/utils/package-json';
+} from '../../../utils/package-json';
 
-export function getNxProjectJsonPlugin(root: string): NxPluginV2 {
-  return {
-    name: 'nx-core-build-project-json-nodes',
-    createNodes: [
-      '{project.json,**/project.json}',
-      (file) => {
-        const json = readJsonFile<ProjectConfiguration>(join(root, file));
-        const project = buildProjectFromProjectJson(json, file);
-        mergePackageJsonConfigurationWithProjectJson(project, root);
-        return {
-          projects: {
-            [project.name]: project,
-          },
-        };
-      },
-    ],
-  };
-}
+export const CreateProjectJsonProjectsPlugin: NxPluginV2 = {
+  name: 'nx-core-build-project-json-nodes',
+  createNodes: [
+    '{project.json,**/project.json}',
+    (file, context) => {
+      const root = context.workspaceRoot;
+      const json = readJsonFile<ProjectConfiguration>(join(root, file));
+      const project = buildProjectFromProjectJson(json, file);
+      mergePackageJsonConfigurationWithProjectJson(project, root);
+      return {
+        projects: {
+          [project.name]: project,
+        },
+      };
+    },
+  ],
+};
 
 export function buildProjectFromProjectJson(
   json: Partial<ProjectConfiguration>,
