--- packages/nx/src/project-graph/build-nodes/index.ts
@@ -1,1 +1,0 @@
-export * from './workspace-projects';
